fix(batch): standardize numeric fields to 2 decimal places + fix feed purchase log
- Change avg_weight_g from SmallInteger to NUMERIC(10,2) in BatchDailyLog model
  and add idempotent migration to upgrade live DB column
- Update DailyLogCreate/Out/Update and BatchSummaryRow schemas: avg_weight_g int → float
- Add ROUND(..., 2) to v_batch_summary view for mortality_pct, total_feed_kg, fcr
- BatchDetailPage: parseInt → parseFloat for avg_weight_g; add step="0.01" to input
- Standardize all .toFixed() calls to 2dp (weight, survival %, margin, feed consumed,
  expense breakdown %, revenue qty_kg, plan phase kg)
- DashboardPage: Feed/Water stat cards unified to toFixed(2)
- FeedPage: Promise.all → Promise.allSettled so one failed endpoint doesn't wipe
  all sections; weekly endpoint guarded when farmId is null
- feed.py: weekly farm_id default 1 → Optional[None] (prevented 403 cascade)
- feed.py: JOIN feed_types → LEFT JOIN to show purchases with deleted feed types

+++ backend/models/models.py
@@ -311,7 +311,7 @@ class BatchDailyLog(Base):
     log_date        = Column(Date, nullable=False)
     current_count   = Column(Integer, nullable=False)
     mortality_count = Column(SmallInteger, nullable=False, default=0)
-    avg_weight_g    = Column(SmallInteger)
+    avg_weight_g    = Column(Numeric(10, 2))
     culls           = Column(SmallInteger, nullable=False, default=0)
     notes           = Column(Text)
     recorded_by     = Column(Integer, ForeignKey("users.id"))
